+++ CHANGES.md
@@ -1,5 +1,7 @@
 ### HEAD
 
+  * Added `Request#apply` and made `Request#call` accept extra arguments
+  * Removed `route` request variable in favor of extra arguments to `Request#apply`
   * Removed support for arrays as a response
 
 ### 0.5.4 / 2013-10-28